requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Bcrypt==0.7.1
PyJWT==1.7.1
Werkzeug==0.15.6
python-dotenv==0.10.3
Pillow==6.1.0
pytesseract==0.3.0
boto3==1.9.215
SQLAlchemy==1.3.8
mysqlclient==1.4.4